aula 3 - finalizada

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -29,6 +36,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+    if (memory.isPublic && memory.userId !== request.user.sub) {
+      throw new Error("Unauthorized");
+    }
     return memory;
   });
 
@@ -44,7 +54,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         isPublic,
         coverUrl,
-        userId: "7745ba55-9f52-47cb-8e7d-692e58f3bb04",
+        userId: request.user.sub,
       },
     });
     return memory;
@@ -62,7 +72,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     });
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+    if (memory.userId !== request.user.sub) {
+      throw new Error("Unauthorized");
+    }
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -76,7 +94,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
       id: z.string().uuid(),
     });
     const { id } = paramsSchema.parse(request.params);
-    const memory = await prisma.memory.delete({
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+    if (memory.userId !== request.user.sub) {
+      throw new Error("Unauthorized");
+    }
+    await prisma.memory.delete({
       where: {
         id,
       },
